v0.1.260528.2: fix compilatiefout app_state.ino — max() type mismatch
max(10UL, (uint32_t)...) faalde: unsigned long en uint32_t zijn
verschillende types op ESP32 (unsigned long vs unsigned int).
Opgelost door beide argumenten naar uint32_t te casten.

+++ BKOS_NUI/app_state.ino
@@ -104,7 +104,7 @@ void state_load() {
         if (key == "wifi_open") wifi_open_auto         = (val.toInt() != 0);
         if (key == "slaap_m")  slaap_modus    = (uint8_t)constrain(val.toInt(), 0, 2);
         if (key == "slaap_t")  slaap_tijd     = (uint32_t)val.toInt();
-        if (key == "slaap_i")  slaap_interval = max(10UL, (uint32_t)val.toInt());
+        if (key == "slaap_i")  slaap_interval = max((uint32_t)10, (uint32_t)val.toInt());
         if (key == "slaap_a")  slaap_attiny   = (val.toInt() != 0);
     }
     f.close();
